I think I'm doing this list pointer thing wrong. Commiting some changes so I can revert back if I need

+++ Entity.hpp
@@ -59,7 +59,7 @@ public:
   bool      idle;
 
   uint8_t   frame;
-  uint8_t   do_frames;
+  int   do_frames;
   uint8_t   framerate;
   uint32_t  last_time;
